Move OpenAI Responses streaming path
